# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "open-moderation-lexicon",
    "description": "Keyword based moderation API built on an open sensitive-word lexicon.\n\nThis service detects keywords. It does not understand meaning, intent or context,\nand `decision` is the result of a configurable policy, not a legal conclusion.\n\nLexicon source: https://github.com/konsheng/Sensitive-lexicon (MIT).",
    "version": "0.1.0",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "TEXT_REQUIRED",
                  "TEXT_NOT_STRING",
                  "TEXT_EMPTY",
                  "TEXT_BLANK",
                  "TEXT_TOO_LONG",
                  "INVALID_JSON",
                  "INVALID_REQUEST",
                  "INVALID_MODE",
                  "INVALID_MASK_CHAR",
                  "UNKNOWN_CATEGORY",
                  "BATCH_EMPTY",
                  "BATCH_TOO_LARGE",
                  "BATCH_ITEM_INVALID",
                  "DUPLICATE_BATCH_ID",
                  "PAYLOAD_TOO_LARGE",
                  "UNSUPPORTED_MEDIA_TYPE",
                  "UNAUTHORIZED",
                  "NOT_FOUND",
                  "METHOD_NOT_ALLOWED",
                  "RATE_LIMITED",
                  "SERVICE_UNAVAILABLE",
                  "LEXICON_LOAD_FAILED",
                  "REQUEST_TIMEOUT",
                  "INTERNAL_ERROR"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "title": "oml:error"
      },
      "def-1": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "term",
          "matchedText",
          "category",
          "severity",
          "matchType",
          "start",
          "end"
        ],
        "properties": {
          "term": {
            "type": "string",
            "description": "The lexicon term that matched."
          },
          "matchedText": {
            "type": "string",
            "description": "Exact substring of the submitted text that matched."
          },
          "category": {
            "type": "string"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high"
            ]
          },
          "matchType": {
            "type": "string",
            "enum": [
              "exact",
              "normalized",
              "fuzzy"
            ],
            "description": "Weakest rule required for this occurrence: exact (literal), normalized (after Unicode/case/traditional folding) or fuzzy (heuristic)."
          },
          "start": {
            "type": "integer",
            "minimum": 0,
            "description": "Inclusive start offset in Unicode code points."
          },
          "end": {
            "type": "integer",
            "minimum": 0,
            "description": "Exclusive end offset in Unicode code points."
          }
        },
        "title": "oml:match"
      },
      "def-2": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "matched",
          "decision",
          "riskLevel",
          "hitCount",
          "uniqueTermCount",
          "categories",
          "matches",
          "lexicon"
        ],
        "properties": {
          "matched": {
            "type": "boolean",
            "description": "True when at least one term matched."
          },
          "decision": {
            "type": "string",
            "enum": [
              "allow",
              "review",
              "block"
            ],
            "description": "Suggested action from the configured policy. Not a legal or regulatory conclusion."
          },
          "riskLevel": {
            "type": "string",
            "enum": [
              "none",
              "low",
              "medium",
              "high"
            ]
          },
          "hitCount": {
            "type": "integer",
            "minimum": 0
          },
          "uniqueTermCount": {
            "type": "integer",
            "minimum": 0
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "matches": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/def-1"
            }
          },
          "maskedText": {
            "type": "string"
          },
          "lexicon": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "version",
              "upstreamCommit"
            ],
            "properties": {
              "version": {
                "type": "string"
              },
              "upstreamCommit": {
                "type": [
                  "null",
                  "string"
                ]
              }
            }
          }
        },
        "title": "oml:moderationResult"
      }
    }
  },
  "paths": {
...
    "/v1/moderate": {
      "post": {
        "summary": "Check one text against the lexicon",
        "tags": [
          "moderation"
        ],
        "description": "Returns keyword matches, a suggested decision and an optional masked text. Offsets are Unicode code point offsets of the submitted text.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 20000,
                    "pattern": "\\S",
                    "description": "Text to check. Must contain at least one non-whitespace character."
                  },
                  "options": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "mode": {
                        "type": "string",
                        "enum": [
                          "exact",
                          "normalized",
                          "fuzzy"
                        ],
                        "description": "Matching mode. Defaults to the server default (normalized)."
                      },
                      "categories": {
                        "type": "array",
                        "minItems": 1,
                        "maxItems": 200,
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 64
                        },
                        "description": "Category ids to match against. Use [\"all\"] for the default selection or [\"*\"] for every loaded category."
                      },
                      "returnMatches": {
                        "type": "boolean",
                        "default": true
                      },
                      "maxMatches": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 100000
                      },
                      "overlap": {
                        "type": "string",
                        "enum": [
                          "all",
                          "leftmost-longest"
                        ]
                      },
                      "mask": {
                        "type": "boolean",
                        "default": false
                      },
                      "maskChar": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 2
                      },
                      "allowlist": {
                        "type": "array",
                        "maxItems": 1000,
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 256
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/def-2"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/v1/moderate/batch": {
      "post": {
        "summary": "Check up to N texts in one request",
        "tags": [
          "moderation"
        ],
        "description": "Each item is processed independently: a malformed item fails without affecting the others. The HTTP status is 200 whenever the envelope itself is valid.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "items"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 100,
                    "items": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "text"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 128
                        },
                        "text": {}
                      }
                    }
                  },
                  "options": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "mode": {
                        "type": "string",
                        "enum": [
                          "exact",
                          "normalized",
                          "fuzzy"
                        ],
                        "description": "Matching mode. Defaults to the server default (normalized)."
                      },
                      "categories": {
                        "type": "array",
                        "minItems": 1,
                        "maxItems": 200,
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 64
                        },
                        "description": "Category ids to match against. Use [\"all\"] for the default selection or [\"*\"] for every loaded category."
                      },
                      "returnMatches": {
                        "type": "boolean",
                        "default": true
                      },
                      "maxMatches": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 100000
                      },
                      "overlap": {
                        "type": "string",
                        "enum": [
                          "all",
                          "leftmost-longest"
                        ]
                      },
                      "mask": {
                        "type": "boolean",
                        "default": false
                      },
                      "maskChar": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 2
                      },
                      "allowlist": {
                        "type": "array",
                        "maxItems": 1000,
                        "items": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 256
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "results",
                        "summary"
                      ],
                      "properties": {
                        "results": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "additionalProperties": false,
                            "required": [
                              "id",
                              "success"
                            ],
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "success": {
                                "type": "boolean"
                              },
                              "data": {
                                "$ref": "#/components/schemas/def-2"
                              },
                              "error": {
                                "type": "object",
                                "additionalProperties": false,
                                "required": [
                                  "code",
                                  "message"
                                ],
                                "properties": {
                                  "code": {
                                    "type": "string",
                                    "enum": [
                                      "TEXT_REQUIRED",
                                      "TEXT_NOT_STRING",
                                      "TEXT_EMPTY",
                                      "TEXT_BLANK",
                                      "TEXT_TOO_LONG",
                                      "INVALID_JSON",
                                      "INVALID_REQUEST",
                                      "INVALID_MODE",
                                      "INVALID_MASK_CHAR",
                                      "UNKNOWN_CATEGORY",
                                      "BATCH_EMPTY",
                                      "BATCH_TOO_LARGE",
                                      "BATCH_ITEM_INVALID",
                                      "DUPLICATE_BATCH_ID",
                                      "PAYLOAD_TOO_LARGE",
                                      "UNSUPPORTED_MEDIA_TYPE",
                                      "UNAUTHORIZED",
                                      "NOT_FOUND",
                                      "METHOD_NOT_ALLOWED",
                                      "RATE_LIMITED",
                                      "SERVICE_UNAVAILABLE",
                                      "LEXICON_LOAD_FAILED",
                                      "REQUEST_TIMEOUT",
                                      "INTERNAL_ERROR"
                                    ]
                                  },
                                  "message": {
                                    "type": "string"
                                  },
                                  "details": {
                                    "type": "object",
                                    "additionalProperties": true
                                  }
                                }
                              }
                            }
                          }
                        },
                        "summary": {
                          "type": "object",
                          "additionalProperties": false,
                          "required": [
                            "total",
                            "succeeded",
                            "failed",
                            "matched"
                          ],
                          "properties": {
                            "total": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "succeeded": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "failed": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "matched": {
                              "type": "integer",
                              "minimum": 0
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "413": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
}